+++ agent-sdk/session/session.go
@@ -283,7 +283,7 @@ type ControllerBinding struct {
 	Label           string         `json:"label,omitempty"`
 	EpochID         string         `json:"epoch_id,omitempty"`
 	RemoteSessionID string         `json:"remote_session_id,omitempty"`
-	ContextSyncSeq  int            `json:"context_sync_seq,omitempty"`
+	ContextSyncSeq  uint64         `json:"context_sync_seq,omitempty"`
 	AttachedAt      time.Time      `json:"attached_at,omitempty"`
 	Source          string         `json:"source,omitempty"`
 }
@@ -302,7 +302,7 @@ type ParticipantBinding struct {
 	// AttachmentGeneration identifies the exact live endpoint instance behind
 	// this durable binding. Conditional detach must match it.
 	AttachmentGeneration string    `json:"attachment_generation,omitempty"`
-	ContextSyncSeq       int       `json:"context_sync_seq,omitempty"`
+	ContextSyncSeq       uint64    `json:"context_sync_seq,omitempty"`
 	AttachedAt           time.Time `json:"attached_at,omitempty"`
 	ControllerRef        string    `json:"controller_ref,omitempty"`
 }
